feat(deploy-infra): integrate CF stack termination protection
Use the new CloudFormation termination protection feature to add some
security to infrastructure stacks.

This works better than the previous "soft" protection afforded by the
tag `yolo:Protected=true`. If yolo looks at a stack created with the old
protection mechanism, update the stack with proper termination
protection.

This patch also removes the `--force` parameter from the `yolo
deploy-infra` command. This means that protected stacks are truly
protected: you can't just blow them away with an extra parameter. In
this case the only way to delete the stack is to use the CloudFormation
console or API to first remove the stack protection and _then_ delete
the stack. This extra step is tedious and that's intentional.