Refactor python builds to use parameterized build image
Parameterize build image to make lambda building faster and
more efficient. Since the image is built as part of the
lambda build process it is no longer necessary to host one
on Docker hub. In addition, this change moves away from
caching wheels and instead caches a zip of all the
dependencies which is more efficient.