client: use overhauled credential/context utils
This is major change. It pretty much touches everything. The general
functionality of yolo doesn't change, but this is a major improvment to
how credentials and runtime context (loaded from various sources) are
handled.

For instance, we no longer are biased towards FAWS credentials; this was
okay for a proof of concept, but being to support basic AWS credentials
in a standard way makes yolo more generally useful.